auto-claude: subtask-2-2 - Run clippy to ensure code quality and no new warnings
Fixed multiple clippy warnings across the codebase:
- Simplified map_or to is_some_and/is_none_or where applicable
- Removed redundant closures
- Removed unnecessary type casts
- Replaced derivable Default implementations with #[derive(Default)]
- Changed ok_or_else to ok_or where appropriate
- Boxed large enum variant to reduce memory usage
- Used strip_prefix instead of manual string slicing
- Converted loop/match to while let pattern
- Collapsed nested if statements
- Added type aliases for complex types
- Suppressed unavoidable await_holding_lock warning with proper justification

+++ crates/at-integrations/src/gitlab/mod.rs
@@ -231,7 +231,7 @@ impl GitLabClient {
     ) -> Result<Vec<GitLabIssue>> {
         if self.is_stub_token() {
             let s = state.unwrap_or("opened");
-            let count = per_page.min(5) as u32;
+            let count = per_page.min(5);
             let issues = (1..=count)
                 .map(|i| Self::stub_issue(project_id, i, s))
                 .collect();
@@ -275,7 +275,7 @@ impl GitLabClient {
     ) -> Result<Vec<GitLabMergeRequest>> {
         if self.is_stub_token() {
             let s = state.unwrap_or("opened");
-            let count = per_page.min(5) as u32;
+            let count = per_page.min(5);
             let mrs = (1..=count)
                 .map(|i| Self::stub_mr(project_id, i, s))
                 .collect();
